Relocate some types and functions
